requirements.txt
pandas==2.2.3
numpy==2.1.3
scikit-learn==1.5.2
joblib==1.4.2

# Backend API
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12
websockets==13.1
sqlalchemy==2.0.36
aiosqlite==0.20.0

# File Monitoring
watchdog==6.0.0
PyPDF2==3.0.1
python-docx==1.1.2

# Utilities
python-dotenv==1.0.1
pydantic==2.10.3
